+++ jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  preset: "@vue/cli-plugin-unit-jest/presets/typescript-and-babel",
-  testMatch: ["<rootDir>/src/**/*.spec.ts"],
-  setupFiles: ["./tests/setup.ts"],
-};